WIP take provider out of lib
