+++ charms/istio-pilot/lib/charms/istio_pilot/v0/istio_gateway_name.py
@@ -119,52 +119,3 @@ class GatewayRequirer(Object):
             "gateway_name": data["gateway_name"],
             "gateway_namespace": data["gateway_namespace"],
         }
-
-
-class GatewayProvider(Object):
-    def __init__(self, charm, relation_name=DEFAULT_RELATION_NAME):
-        super().__init__(charm, relation_name)
-
-        from lightkube.core.client import Client
-        from lightkube.generic_resource import create_namespaced_resource
-
-        self.charm = charm
-        self.lightkube_client = Client(namespace=self.model.name, field_manager="lightkube")
-        self.gateway_class = create_namespaced_resource(
-            group="networking.istio.io",
-            version="v1beta1",
-            kind="Gateway",
-            plural="gateways",
-            verbs=None,
-        )
-        self.framework.observe(
-            charm.on[relation_name].relation_changed, self._on_gateway_relation_changed
-        )
-        self.framework.observe(charm.on.config_changed, self._on_gateway_relation_changed)
-        self.framework.observe(charm.on.update_status, self._on_gateway_relation_changed)
-
-    def _validate_gateway_exists(self):
-        from lightkube.core.exceptions import ApiError
-
-        try:
-            response = self.lightkube_client.get(
-                self.gateway_class, self.model.config['default-gateway'], namespace=self.model.name
-            )
-            return True
-        except ApiError as error:
-            logger.error(str(error))
-
-        return False
-
-    def _on_gateway_relation_changed(self, event):
-        if self.model.unit.is_leader():
-            relations = self.model.relations["gateway"]
-            for relation in relations:
-                relation.data[self.charm.app].update(
-                    {
-                        "gateway_name": self.model.config["default-gateway"]
-                        if self._validate_gateway_exists()
-                        else "",
-                        "gateway_namespace": self.model.name,
-                    }
-                )
